glm 4v 1st sdp for vision (#12904)
* glm4v 1st sdp

* update glm4v example

* meet code review

* fix style

+++ python/llm/example/GPU/HuggingFace/Multimodal/glm-4v/generate.py
@@ -61,7 +61,7 @@ if __name__ == '__main__':
                                                  trust_remote_code=True,
                                                  use_cache=True,
                                                  model_hub=model_hub)
-    model = model.half().to('xpu')
+    model = model.to('xpu')
     
     tokenizer = AutoTokenizer.from_pretrained(model_path, trust_remote_code=True)
